first_deliverable/requirements.txt
ament_index_python==1.4.0
builtin_interfaces==1.2.1
cmake_build_extension==0.5.1
cv_bridge==3.2.1
gazebo_msgs==3.7.0
geometry_msgs==4.2.3
image_geometry==3.2.1
launch_ros==0.19.6
launch_testing==1.0.4
message_filters==4.3.3
nav2_common==1.1.12
numpy==1.21.5
pytest==6.2.5
PyYAML==5.4.1
PyYAML==6.0.1
rclpy==3.3.10
Requests==2.31.0
rosidl_parser==3.1.5
sensor_msgs==4.2.3
setuptools==58.2.0
setuptools==59.6.0
std_msgs==4.2.3
std_srvs==4.2.3
stereo_msgs==4.2.3
tf2_ros_py==0.25.4
tf_transformations==1.0.0
turtlesim==1.4.2
visualization_msgs==4.2.3
